requirements.txt
fastapi
uvicorn
groq
faiss-cpu
sentence-transformers
PyPDF2
python-multipart
pydantic
